requirements.txt
pandas==2.3.3
numpy==2.3.3
requests==2.31.0
jinja2==3.1.2

apache-airflow==2.9.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0    
